experimental: remove broken runtime.LockOSThread behavior -- simplified txns
Write transactions are run similar to in BoltDB, by passing a
function/closure to a goroutine that is locked on a thread. There are
still possibilities of undefined behavior if the function leaks the
transaction to another goroutine.  But accidentally relying on undefined
behavior should be far less common now.

The example tests have been rewritten using the new methods but most
other tests have not (and are probably technically broken now).  Despite
this all tests still pass.