initial attempt at supporting mdb_reader_list
Calls are managed using global int64 context object.  Callback functions
for each context are stored in a map guarded by an RWMutex.  In order to
satisfy the C.MDB_msg_func type a proxy written in C was required
(AFAIK).

It's possible that this file could be eliminated through static
inlining.  The documentation and community knowledge on the matter is
spotty.

https://golang.org/cmd/cgo/
https://github.com/golang/go/wiki/cgo
http://blog.golang.org/c-go-cgo