some helpers for (un)wrapping unsigned integer types and lmdb.mdbVal
When creating an lmdb.mdbVal it's critical that the caller maintain a
reference to the integer for the duration of Txn./Cursor.Put to avoid
invalid stack memory address dereferencing.

Wrapping may need to be managed at the transaction level.  It would
probably place too much burden on the user otherwise.  However, adding
Txn methods to support all variations of {byte, uint, uint64}
combinations would mean adding a lot of methods (8 -- 1 exists).

The portable alternative to MDB_INTEGERKEY/MDB_INTEGERDUP is to use an
explicit bit width and byte encoding (probably BE for sort order).  The
disadvantage is probably is comparision time.  Benchmarks need to be
written to quantify the difference is speed.

- What are the right kinds of benchmarks?  Random Reads?  Range Reads?
  Random Writes?  Table Scan?