Docs: replace exampleopendbi_test.go with several small examples
This test was the most difficult to replace because it described error
conditions to some success.  Strict code examples do this less well.
The replacement examples seem to do a reasonably good job due to their
brevity.

Examples include more general enforcement of the idea that DBIs should
be held for the duration of an enviornment's use.

Reference to flag and errno values in the package was made consistent so
that comment text does not use an "lmdb." prefix on names.

Closes #2.