lmdb: Cursor.Put* and Cursor.Get use safe cgo proxy functions
MDB_val data is passed separately (and exploded) to avoid the violating
the constraint the Go pointer arguments not point to memory containing
Go pointers.

Go pointers to empty MDB_val structs are passed as well so they may be
used to contain the result from mdb_cursor_get or mdb_cursor_put.  This
bypasses the cgo pointer argument restrictions because the Go memory
pointed to by the arguments contains only nil pointers.