lmdb: simplify internals for Env.ReaderList to use a single map
The previous use of two maps to hold the func and error of a context was
complicated.  Using a single map containing a struct pointer is simpler.
It may not always be as efficient but the effect should be negligible
and the Env.ReaderList function is not generally critical for
performance.