lmdbscan: fix panic in Scanner.Scan after Txn.OpenCursor failed
Fixes #21

Method Scanner.Scan, checks the underlying cursor before calling
Cursor.Get.  This allows a Scanner.Err to catch an error opening the
cursor even if called at the end of the transaction.

If the Scanner.Scan method is called after the scanner has been closed
then a Scanner.Err is set to return an error indicating that the scanner
is closed.