all: internal test package lmdbtest that performs common test tasks
Fixes #24

Logic that was either repeated or likely to be repeated was subsumed
into the new package.  The exp/lmdbscan and exp/lmdbsync packages have
had their tests converted to use the new package.

It's unclear if the core lmdb package can make use  of these helper
functions.  It is possible that, if a "black box" test style is used,
the internal package can be successfully imported.  But it is likely
that setup/teardown functions will need to remain defined in that
package.