lmdb: the Txn finalizer logs that an unterminated txn was found
An unterminated Txn is serious business.  And to terminate it silently
is a disservice to the user.  Once the user is in the situation that
they have unterminated txns it will be handled but they must be alerted
to this fact.

It's unclear if there should be a mode where an unterminated Txn cause a
runtime panic.  Though, allowing the user to set the log func is
probably a good idea eventually.