lmdb: special error handling for windows (vs unix)
The errno values returned by LMDB are not standard Windows error codes
and must be translated to the invented values in the syscall package for
the lmdb package API to make sense.  This may not be the permanent
solution for error handling on Windows.  And, a future version of LMDB
will make a change to use only custom error codes.