lmdbsync: changes to MapResizedHandler delays and defaults
A DelayFunc type is defined to describe the interface for
MapResizedHandler.  The delay function is called before each call to
Env.SetMapSize including the first, when its argument is zero.  This is
an incompatible change for lmdbsync.

Additionally, the default values for MapResizedHandler arguments have
been renamed so that they more clearly reflect their purpose.  Their
previous names made no sense to users as they corresponded to field
names for unexported structs.