as an experiment, switch void* cgo arguments with char*
It was discovered in #63 that void* arguments cause allocations in cgo
and extra time checking arguments as opposed to char* arguments.

The issue golang/go#14387 claims that it will fix this problem but I am
going to keep this patch around for testing against tip and later
merging if necessary.