docs: clarify limitations on the number of open databases for LMDB
There was one article I read (written by Ben B Johnson I believe) which
recomended storing an unbounded (and fast growing) number of buckets in
a BoltDB database (there may be more such articles).  It should be clear
to potential users comparing the databases that LMDB is not designed to
work that way.