lmdb: Txn.OpenDBI() just proxies the dbi from C
It is most transparent to simply use whatever DBI value is set (or not
set as it were) by mdb_open_dbi.  This causes OpenDBI and OpenRoot to
return the zero DBI.  This is not necessarily correct but it is
backwards compatible.  And because the package does not promote ignoring
error values it is not too unsafe.  If there is a real need to break API
compatibility in the future then this decision could be revisited.