lmdb: Rename MultiUint[ptr] to Uint[ptr]Multi
This will provide better grouping of topics in godoc.  Uint stuff will
all be together and Uintptr stuff will all be together.

The byte-related Value stuff won't really be together, but that is a
tragedy of evolution.  Maybe for lmdb-go 2.0 this part of the Value and
FixedPage interfaces can be made consistent for all data types.