lmdb: Reorganize definitions into new file data.go
This is where the Data interface is defined as well as the generic
BytesData type.  This is also where the FixedPage interface is defined
because it seems like the most appropriate place, though the generic
Multi type has been left if val.go, mostly for historical reasons.  It
may be moved later.