lmdb: Rename FixedPage to FixedMultiple change Multi constructors
The name FixedPage never sat right with me.  FixedMultiple sounds more
correct.  It is better that the implementations are named *Multi.

I don't particularly care for the name Multi, but it has to stick for
now because of backwards compatability (the Multi type has existed for a
very long time).

WrapMulti is particularly bad, in my opinion (which is also stuck).  I
have chosen to call the typed functions UintMultiple and
UintptrMultiple.  I may change these functions back for consistency
though.