lmdb: Txn objects do not have their finalizers cleared in Txn.Reset
Relates to #104.  This behavior was undocumented, hard to detect, and
harmful in any situation where it attempted to help the programmer.

The finalizer now properly detects when it would have otherwise caused a
panic by trying to abort the transaction after its Env had been closed.
The Txn uses a sync.RWMutex to make sure that Env does not get closed
while the Txn is being aborted.

A Pooled field has been added to the Txn type that allows disabling of
the Txn's warning when its finalizer is called.  The idea being that an
application or package making use of a sync.Pool may reasonably manage
this field and safely provide readonly transaction pooling.