lmdb: Method Txn.RunOp allows other packages to implement managed Txns
It was possible for packages to wrap calls to Env.Update, Env.View, etc.
But it wasn't really possible for packages to implement managed
transactions from scratch.

The new Txn.RunOp method makes this possible.  For now, it is used to
create all managed transactions, though this may prove to be less
efficient than is really desirable in the end.