lmdbpool: New experimental package
While recent commits have allowed safe pooling of transactions in
theory, it is not trivial to pool transactions.  The new
lmdbpool.TxnPool type is an experimental reference implementation to
transaction pooling.  If the benefits of lmdbpool can be vetted more
thoroughly, its API can be stabilized and likely incorporated into the
lmdb package itself.

Until pooling is directly integrated into the lmdb package the TxnPool
type provides its own implementations of the Update and View methods
which integrate directly into the TxnPool and are the recommended way to
create all transactions when pooling is done.