lmdb: rework BenchmarkScan to show Update with no contention is faster
It is not obvious that an update transaction will be faster than a
Readonly transaction, even when renewing both the Txn and Cursor object.
This benchmark rework is the first in a series that I hope to make, and
that I hope will continue to bring meaningful insight into LMDB
performance, unlike many of the existing benchmarks.