lmdb: optimize Txn.ID() so that successive calls do not call cgo
A transaction's ID typically does not change.  This is true up until the
transaction in terminated (not including being reset -- renewed
transaction retain their old id).  So the value returned Txn.ID() may be
cached up until Txn.Commit() or Txn.Abort() is called.

An internal method Txn.clearTxn() was added to make sure that when the
C.MDB_txn object in a Txn is zeroed any cached ID is also zeroed.

This change should be transparent to normal users.  The Txn.ID() method
is typically not called often, but it appears that Txn pooling can
benefit from checking Txn.ID() multiple times (as the Txn enters/exits
the pool)