lmdbpool: Add UpdateHandling type to tell TxnPool how to handle updates
It's just not clear right now what the correct behavior is, so it is
configurable.  Ideally, a "good enough" value can be determined and made
the default, potentially disallowing overriding all together.

If the behavior can't be overridden in TxnPool then, presuming this is
incorporated into the lmdb package the default TxnPool used by an
lmdb.Env must allow overrides (e.g. using an interface).