lmdbpool: TxnPool.CommitID handles idle transactions in the pool
A new goroutine is spawned for each update.  This may turn out to be too
costly.  We want updates to complete as fast as possible and handling
the idle transactions in that goroutine is not practical.  So it really
just comes down to doing things this way or using a permanent working
goroutine that would require more management/bookkeeping.